# swagger.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/subgrupo-detalle/detalle/{id}": {
            "get": {
                "operationId": "SubgrupoDetalleController_getDetalle",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "subgrupo-detalle"
                ]
            }
        },
...
        "/seguimiento": {
            "post": {
                "operationId": "SeguimientoController_post",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SeguimientoDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "seguimiento"
                ]
            },
            "get": {
                "operationId": "SeguimientoController_getAll",
                "parameters": [
                    {
                        "name": "query",
                        "required": false,
                        "in": "query",
                        "description": "query - Filter. e.g. col1:v1,col2:v2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fields",
                        "required": false,
                        "in": "query",
                        "description": "fields - Fields returned. e.g. col1,col2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sortby",
                        "required": false,
                        "in": "query",
                        "description": "sortby - Sorted-by fields. e.g. col1,col2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order",
                        "required": false,
                        "in": "query",
                        "description": "order - Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "description": "limit - Limit the size of result set. Must be an integer",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "offset",
                        "required": false,
                        "in": "query",
                        "description": "offset - Start position of result set. Must be an integer",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "populate",
                        "required": false,
                        "in": "query",
                        "description": "populate - show subqueries. Must be a boolean",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "seguimiento"
                ]
            }
        },
        "/seguimiento/{id}": {
            "get": {
                "operationId": "SeguimientoController_get",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "seguimiento"
                ]
            },
            "put": {
                "operationId": "SeguimientoController_put",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SeguimientoDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "seguimiento"
                ]
            },
            "delete": {
                "operationId": "SeguimientoController_delete",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "seguimiento"
                ]
            }
        },
...
        "/seguimiento-detalle": {
            "post": {
                "operationId": "SeguimientoDetalleController_post",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SeguimientoDetalleDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "seguimiento-detalle"
                ]
            },
            "get": {
                "operationId": "SeguimientoDetalleController_getAll",
                "parameters": [
                    {
                        "name": "query",
                        "required": false,
                        "in": "query",
                        "description": "query - Filter. e.g. col1:v1,col2:v2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fields",
                        "required": false,
                        "in": "query",
                        "description": "fields - Fields returned. e.g. col1,col2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sortby",
                        "required": false,
                        "in": "query",
                        "description": "sortby - Sorted-by fields. e.g. col1,col2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order",
                        "required": false,
                        "in": "query",
                        "description": "order - Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "description": "limit - Limit the size of result set. Must be an integer",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "offset",
                        "required": false,
                        "in": "query",
                        "description": "offset - Start position of result set. Must be an integer",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "populate",
                        "required": false,
                        "in": "query",
                        "description": "populate - show subqueries. Must be a boolean",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "seguimiento-detalle"
                ]
            }
        },
        "/seguimiento-detalle/{id}": {
            "get": {
                "operationId": "SeguimientoDetalleController_get",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "seguimiento-detalle"
                ]
            },
            "put": {
                "operationId": "SeguimientoDetalleController_put",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SeguimientoDetalleDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "seguimiento-detalle"
                ]
            },
            "delete": {
                "operationId": "SeguimientoDetalleController_delete",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "seguimiento-detalle"
                ]
            }
        },
...
    },
    "info": {
        "title": "plan_crud",
        "description": "API CRUD para el registro de planes para el cliente de planeacion",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "schemas": {
...
            "SeguimientoDto": {
                "type": "object",
                "properties": {
                    "nombre": {
                        "type": "string"
                    },
                    "descripcion": {
                        "type": "string"
                    },
                    "plan_id": {
                        "type": "string"
                    },
                    "dato": {
                        "type": "string"
                    },
                    "tipo_seguimiento_id": {
                        "type": "string"
                    },
                    "estado_seguimiento_id": {
                        "type": "string"
                    },
                    "periodo_seguimiento_id": {
                        "type": "string"
                    },
                    "activo": {
                        "type": "boolean"
                    },
                    "fecha_inicio": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "fecha_fin": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "fecha_creacion": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "fecha_modificacion": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "nombre",
                    "descripcion",
                    "plan_id",
                    "dato",
                    "tipo_seguimiento_id",
                    "estado_seguimiento_id",
                    "periodo_seguimiento_id",
                    "activo",
                    "fecha_inicio",
                    "fecha_fin",
                    "fecha_creacion",
                    "fecha_modificacion"
                ]
            },
...
            "SeguimientoDetalleDto": {
                "type": "object",
                "properties": {
                    "cualitativo": {
                        "type": "string"
                    },
                    "cuantitativo": {
                        "type": "string"
                    },
                    "estado": {
                        "type": "string"
                    },
                    "evidencia": {
                        "type": "string"
                    },
                    "informacion": {
                        "type": "string"
                    },
                    "activo": {
                        "type": "boolean"
                    },
                    "fecha_creacion": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "fecha_modificacion": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "cualitativo",
                    "cuantitativo",
                    "estado",
                    "evidencia",
                    "informacion",
                    "activo",
                    "fecha_creacion",
                    "fecha_modificacion"
                ]
            },
...
        }
    }
}